refactoring - moving enitites to be deleted to a separate package
