+++ src/main/java/com/isa/morswiny/comparators/DateComparator.java
@@ -1,6 +1,6 @@
 package com.isa.morswiny.comparators;
 
-import com.isa.morswiny.events.Event;
+import com.isa.morswiny.toBeDeleted.Event;
 import java.util.Comparator;
 
 public class DateComparator implements Comparator<Event> {
