+++ src/main/java/com/isa/morswiny/eventsDao/EventCRUDRepositoryInterface.java
@@ -1,9 +1,8 @@
 package com.isa.morswiny.eventsDao;
 
-import com.isa.morswiny.events.Event;
+import com.isa.morswiny.toBeDeleted.Event;
 
 import javax.ejb.Local;
-import javax.inject.Inject;
 import java.util.List;
 
 @Local
