+++ src/main/java/com/isa/morswiny/eventsDao/EventSearchRepository.java
@@ -1,7 +1,7 @@
 package com.isa.morswiny.eventsDao;
 
-import com.isa.morswiny.events.Event;
-import com.isa.morswiny.repository.EventRepository;
+import com.isa.morswiny.toBeDeleted.Event;
+
 import javax.enterprise.context.SessionScoped;
 import javax.inject.Inject;
 import java.io.Serializable;
