+++ src/main/java/com/isa/morswiny/repository/JsonEventDataLoad.java
@@ -3,7 +3,7 @@ package com.isa.morswiny.repository;
 
 import com.google.gson.Gson;
 
-import com.isa.morswiny.events.Event;
+import com.isa.morswiny.toBeDeleted.Event;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
