+++ src/main/java/com/isa/morswiny/servlets/SearchEventsServletBeta.java
@@ -1,6 +1,6 @@
 package com.isa.morswiny.servlets;
 
-import com.isa.morswiny.events.Event;
+import com.isa.morswiny.toBeDeleted.Event;
 import com.isa.morswiny.eventsDao.EventCRUDRepositoryInterface;
 import com.isa.morswiny.eventsDao.EventSearchRepositoryInterface;
 import com.isa.morswiny.freemarker.TemplateProvider;
