+++ src/main/java/com/isa/morswiny/toBeDeleted/Attachment.java
@@ -1,27 +1,19 @@
-package com.isa.morswiny.events;
+package com.isa.morswiny.toBeDeleted;
 
 import javax.persistence.*;
 
-@Entity
-@Table(name="attachment")
-public class AttachmentEntity {
+//@Entity
+//@Table(name="attachment")
+public class Attachment {
     public Integer getAttachmentId() {
         return attachmentId;
     }
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
     private Integer attachmentId;
     private String fileName;
-    @ManyToOne
-    @JoinColumn(name = "event_Id", referencedColumnName = "eventId")
-    private EventEntity event;
-
-    public AttachmentEntity(String fileName) {
+    public Attachment(String fileName) {
         this.fileName = fileName;
     }
-
-    public AttachmentEntity() {
+    public Attachment() {
     }
 
     @Override
